chore(swingset): move buildDispatch out to util.js
