+++ packages/SwingSet/test/util.js
@@ -21,3 +21,41 @@ export function checkKT(t, kernel, expected) {
   expected.sort(compareArraysOfStrings);
   t.deepEqual(got, expected);
 }
+
+
+export function buildDispatch(onDispatchCallback = undefined) {
+  const log = [];
+
+  const dispatch = {
+    deliver(targetSlot, method, args, resultSlot) {
+      const d = { type: 'deliver', targetSlot, method, args, resultSlot };
+      log.push(d);
+      if (onDispatchCallback) {
+        onDispatchCallback(d);
+      }
+    },
+    notifyFulfillToPresence(promiseID, slot) {
+      const d = { type: 'notifyFulfillToPresence', promiseID, slot };
+      log.push(d);
+      if (onDispatchCallback) {
+        onDispatchCallback(d);
+      }
+    },
+    notifyFulfillToData(promiseID, data) {
+      const d = { type: 'notifyFulfillToData', promiseID, data };
+      log.push(d);
+      if (onDispatchCallback) {
+        onDispatchCallback(d);
+      }
+    },
+    notifyReject(promiseID, data) {
+      const d = { type: 'notifyReject', promiseID, data };
+      log.push(d);
+      if (onDispatchCallback) {
+        onDispatchCallback(d);
+      }
+    },
+  };
+
+  return { log, dispatch };
+}
